refactor: remove production mocks; use fixtures in tests/scripts
